Track only motion
The fact that the buttons right above the touchpad actually report
events for the trackpoint puts a wrench into the current device
tracking idea. I always use those buttons together with the touchpad,
so I constantly change the current device between touchpad and
trackpoint, which ruins the experience.

To overcome this problem, use only motion events for the current
tracking, eliminating the problem, but making this tracking less
generally useful. It is really tailored towards overlay scrolling
now.