cssselector: Reintroduce GOT_MATCH
I removed it in 14f5ce710856708dc80fcc98cb5eb2ba257f34ad because I
thought it was unnecessary, but it wasn't. When we build a tree like
this:

.matches ─┬─ .doesntmatch
          └─ .alsodoesntmatch

We would get the changes for the .matches part returned. This is however
only right if that node of the tree contains results. If results only
exist with the child nodes (all of which don't match), then this part
should not match either.