roundedbox: Treat path segments differently
If a side of the box is 0px wide, make the corners owned by the adjacent
sides. This avoids spilling over of unwanted colors from the 0-width
side into the corner.

This is the fix that should actually have been part of
7e5ac3a58d0362080c4eb7c9b68ead99674c7b87 but my amazing amending skills
got rid of it when amending the test.