dnd: Allow programmatic drag cancellation
Add a gtk_drag_cancel() function that can be used on the
source side to cancel an ongoing drag operation. This can
be useful, e.g. if the data that is being dragged becomes
unavailable.

gtk_drag_cancel() will become public API in 3.16; we keep it
private here so we can use it from GtkNotebook.